pyproject.toml
[project]
name = "condor"
version = "0.1.0"
description = "Async Frigate ZMQ remote object detector server"
requires-python = ">=3.12, <3.13"
dependencies = [
    "pyzmq>=26.0",
    "aiofiles>=24.0",
    "numpy>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "onnxruntime>=1.19",
    "opencv-python-headless>=4.10",
]

[project.optional-dependencies]
openvino = [
    # Replace onnxruntime with the OpenVINO-bundled build.
    # Install with: uv pip install onnxruntime-openvino --no-deps
    # then: uv pip uninstall onnxruntime
    "onnxruntime-openvino",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
condor = "condor.server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["condor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
